% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMetricFunctions.R
\name{calculate_coverage_rate}
\alias{calculate_coverage_rate}
\title{Calculate Coverage Rate of Prediction Intervals for Time Series}
\usage{
calculate_coverage_rate(actual_values, lower_bounds, upper_bounds)
}
\arguments{
\item{actual_values}{A numeric vector or a time series object (`ts`) containing the actual observed values.}

\item{lower_bounds}{A numeric vector or a time series object (`ts`) containing the lower bounds of the prediction intervals.}

\item{upper_bounds}{A numeric vector or a time series object (`ts`) containing the upper bounds of the prediction intervals.}
}
\value{
A numeric value between 0 and 1 representing the proportion of observations within the prediction intervals.
}
\description{
This function computes the coverage rate of prediction intervals for a given time series.
}
\examples{
# Example usage with numeric vectors
actual_values <- ts(c(100, 105, 110, 95, 120), start = c(2023, 1), frequency = 12)
lower_bounds <- ts(c(90, 100, 105, 85, 110), start = c(2023, 1), frequency = 12)
upper_bounds <- ts(c(110, 115, 120, 105, 130), start = c(2023, 1), frequency = 12)
calculate_coverage_rate(actual_values, lower_bounds, upper_bounds)
}
