% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMetricFunctions.R
\name{msis}
\alias{msis}
\title{Calculate Mean Scaled Interval Score (MSIS)}
\usage{
msis(actual, lower, upper, scale, alpha = 0.05)
}
\arguments{
\item{actual}{A numeric vector of actual values.}

\item{lower}{A numeric vector of lower bounds of the prediction intervals.}

\item{upper}{A numeric vector of upper bounds of the prediction intervals.}

\item{scale}{A numeric value for scaling, typically based on the mean absolute 
error of a naive forecast over a training period.}

\item{alpha}{A numeric value for the significance level of the interval (default is 0.05).}
}
\value{
A single numeric value representing the MSIS.
}
\description{
Computes the MSIS for prediction intervals, measuring the accuracy and precision 
of forecast intervals while considering scaling based on past errors.
}
\examples{
actual <- c(100, 110, 90)
lower <- c(95, 105, 85)
upper <- c(105, 115, 95)
scale <- 10
msis(actual, lower, upper, scale)
}
